fix: Make Nyx kernel compile on bare-metal x86_64 target
Major compilation fixes for Rust nightly 1.93:
- Fix CPUID inline asm to save/restore rbx (LLVM reserved register)
- Add explicit unsafe blocks for deny(unsafe_op_in_unsafe_fn)
- Add Send implementations for heap allocator types (NonNull<T>)
- Fix SocketAddr/IpAddr traits (PartialOrd, Ord, Display)
- Add Clone for AddressSpace and CSpace
- Fix Thread/Process struct field mismatches
- Fix ThreadState::Blocked variant (requires BlockReason)
- Fix DeadlineQueue::enqueue to use DeadlineEntry
- Fix IpcError pattern matching for InternalError
- Fix GDT inline asm label (avoid binary-like names)
- Make map_page public for checkpoint restoration
- Update panic handler for new PanicMessage API
- Fix Rights::from_bits_truncate to use u64

+++ kernel/src/syscall.rs
@@ -8,7 +8,7 @@ use crate::ipc;
 use crate::mem::{VirtAddr, PAGE_SIZE};
 use crate::process::{ProcessId, SpawnArgs, SpawnError};
 use crate::sched::{SchedClass, ThreadState, BlockReason};
-use alloc::string::String;
+use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use core::time::Duration;
 
@@ -346,7 +346,7 @@ fn handle_poll(regs: &mut SyscallRegs) -> Result<u64, SyscallError> {
 
 fn handle_cap_derive(regs: &mut SyscallRegs) -> Result<u64, SyscallError> {
     let src_cap = regs.arg0;
-    let new_rights = Rights::from_bits_truncate(regs.arg1 as u32);
+    let new_rights = Rights::from_bits_truncate(regs.arg1);
 
     match crate::cap::derive(ObjectId::from_raw(src_cap), new_rights) {
         Ok(cap) => Ok(cap.object_id.as_u64()),
